Finalised Code
To deliver source code and binary APK file for a sample weather
application running on Android OS with the following minimal functional
requirements.

-	The app obtains the current GPS location of the device, which invokes
a web-service request to fetch the weather for identified location.
-	When a response has been received by the calling app thread the
returned JSON data is parsed and de-serialized into an appropriate data
model.
-	This data is displayed on the app UI using a Google Map view.
-	The app allows the display of a map with the user’s current location
and weather.
-	The user is able to interact with the map (zoom, pan, etc.), and
after clicking on the map, all the relevant weather data for the
user-clicked location is displayed in an overlay on the map.
-	Data caching is handled by a worker thread, which provides data
storage functionality.
-	The app can be used in portrait and landscape mode. Providing an
intuitive UI adhering to Android conventions.